# service-template/cli/cmd/template/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger tools for {{PROVIDER}} integrator",
        "title": "{{SERVICE_TITLE}}",
        "contact": {
            "name": "@supplyTeam"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/{{BASE_PATH}}/booking": {
            "post": {
                "description": "endpoint for booking process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prime-booking"
                ],
                "summary": "Booking endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "TIKETCOM",
                        "description": "Authentication header",
                        "name": "storeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "WEB",
                        "description": "Authentication header",
                        "name": "channelId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "321321321",
                        "description": "Authentication header",
                        "name": "requestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "guest",
                        "description": "Authentication header",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "GATEWAY",
                        "description": "Authentication header",
                        "name": "serviceId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "resellerId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "identity",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/book.IntegratorBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/book.BookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/{{BASE_PATH}}/cancel-booking": {
            "post": {
                "description": "Endpoint for Cancel Booking process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prime-booking"
                ],
                "summary": "Cancel-Booking endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "TIKETCOM",
                        "description": "Authentication header",
                        "name": "storeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "WEB",
                        "description": "Authentication header",
                        "name": "channelId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "321321321",
                        "description": "Authentication header",
                        "name": "requestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "guest",
                        "description": "Authentication header",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "GATEWAY",
                        "description": "Authentication header",
                        "name": "serviceId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "resellerId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "identity",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/cancel_book.IntegratorCancelBookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/cancelbook.CancelBookResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
...
        "/{{BASE_PATH}}/issued": {
            "post": {
                "description": "Endpoint for issued process",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "prime-booking"
                ],
                "summary": "Issued endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "default": "TIKETCOM",
                        "description": "Authentication header",
                        "name": "storeId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "WEB",
                        "description": "Authentication header",
                        "name": "channelId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "321321321",
                        "description": "Authentication header",
                        "name": "requestId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "guest",
                        "description": "Authentication header",
                        "name": "username",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "GATEWAY",
                        "description": "Authentication header",
                        "name": "serviceId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "resellerId",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Authentication header",
                        "name": "identity",
                        "in": "header"
                    },
                    {
                        "description": "Request Body",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/issued.IntegratorIssuedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/issued.IssuedResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "book.BookDetail": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "bookingCode": {
                    "type": "string"
                },
                "bookingDate": {
                    "type": "string"
                },
                "bookingExpired": {
                    "type": "string"
                },
                "corporateCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "distributionType": {
                    "type": "string"
                },
                "invoiceNumber": {
                    "type": "string"
                },
                "itineraries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/book.BookItinerary"
                    }
                },
                "logId": {
                    "type": "string"
                },
                "originalBookingFare": {
                    "$ref": "#/definitions/book.BookFare"
                },
                "originalCurrency": {
                    "type": "string"
                },
                "passengerFareInfo": {
                    "$ref": "#/definitions/book.PassengerFareInfo"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
...
        "book.BookResponse": {
            "type": "object",
            "properties": {
                "bookDetail": {
                    "$ref": "#/definitions/book.BookDetail"
                },
                "integratorBookRequest": {
                    "$ref": "#/definitions/book.IntegratorBookRequest"
                },
                "mandatoryRequest": {
                    "$ref": "#/definitions/common.MandatoryRequest"
                }
            }
        },
...
        "book.Contact": {
            "type": "object",
            "properties": {
                "dob": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailAgent": {
                    "type": "string"
                },
                "emailPassenger": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "phoneCode": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profileId": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "book.IntegratorBookRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "adult": {
                    "type": "integer"
                },
                "child": {
                    "type": "integer"
                },
                "commission": {
                    "type": "number"
                },
                "contact": {
                    "$ref": "#/definitions/book.Contact"
                },
                "corporate": {
                    "$ref": "#/definitions/fare.Corporate"
                },
                "countryId": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "distributionType": {
                    "type": "string"
                },
                "fareSellKey": {
                    "type": "string"
                },
                "flightSelect": {
                    "type": "string"
                },
                "infant": {
                    "type": "integer"
                },
                "itineraries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/book.Itinerary"
                    }
                },
                "journeySellKey": {
                    "type": "string"
                },
                "maxBookRetry": {
                    "type": "integer"
                },
                "multiFare": {
                    "type": "boolean"
                },
                "paxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/book.Pax"
                    }
                },
                "promoCode": {
                    "type": "string"
                },
                "sessionGroup": {
                    "type": "string"
                },
                "sessionGroupType": {
                    "type": "string"
                },
                "viaSelect": {
                    "type": "string"
                }
            }
        },
        "book.Itinerary": {
            "type": "object",
            "properties": {
                "airline": {
                    "type": "string"
                },
                "arrival": {
                    "type": "string"
                },
                "blacklistMap": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "brandedFare": {
                    "$ref": "#/definitions/fare.BrandedFare"
                },
                "cabinClass": {
                    "$ref": "#/definitions/enum.CabinType"
                },
                "connecting": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "departure": {
                    "type": "string"
                },
                "international": {
                    "type": "integer"
                },
                "schedules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/book.Schedule"
                    }
                }
            }
        },
...
        "book.Pax": {
            "type": "object",
            "properties": {
                "baggage": {
                    "$ref": "#/definitions/book.Baggage"
                },
                "dob": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "frequentFlyer": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "identityNumber": {
                    "$ref": "#/definitions/book.IdentityNumber"
                },
                "insurance": {
                    "type": "boolean"
                },
                "lastName": {
                    "type": "string"
                },
                "meals": {
                    "$ref": "#/definitions/book.Meals"
                },
                "nationality": {
                    "type": "string"
                },
                "originalFirstName": {
                    "type": "string"
                },
                "originalLastName": {
                    "type": "string"
                },
                "passport": {
                    "$ref": "#/definitions/book.Passport"
                },
                "paxNumber": {
                    "type": "integer"
                },
                "phoneCode": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profileId": {
                    "type": "string"
                },
                "seats": {
                    "$ref": "#/definitions/book.Seats"
                },
                "title": {
                    "type": "string"
                },
                "travelPurpose": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "cancel_book.IntegratorCancelBookRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "bookingCode": {
                    "type": "string"
                },
                "distributionType": {
                    "type": "string"
                },
                "sessionGroup": {
                    "type": "string"
                },
                "sessionGroupType": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
        "cancelbook.CancelBookResponse": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "bookingCode": {
                    "type": "string"
                },
                "distributionType": {
                    "type": "string"
                },
                "integratorCancelBookRequest": {
                    "$ref": "#/definitions/cancel_book.IntegratorCancelBookRequest"
                },
                "mandatoryRequest": {
                    "$ref": "#/definitions/common.MandatoryRequest"
                }
            }
        },
        "common.MandatoryRequest": {
            "type": "object",
            "required": [
                "channelId",
                "requestId",
                "serviceId",
                "storeId",
                "username"
            ],
            "properties": {
                "channelId": {
                    "type": "string"
                },
                "identity": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "resellerId": {
                    "type": "string"
                },
                "serviceId": {
                    "type": "string"
                },
                "storeId": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "fare.Account": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "fare.Corporate": {
            "type": "object",
            "properties": {
                "airlines": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "corporateCodes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "corporateId": {
                    "type": "integer"
                }
            }
        },
...
        "issued.IntegratorIssuedRequest": {
            "type": "object",
            "properties": {
                "account": {
                    "$ref": "#/definitions/fare.Account"
                },
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "balanceDue": {
                    "type": "number"
                },
                "bookingCode": {
                    "type": "string"
                },
                "bookingDate": {
                    "type": "string"
                },
                "commission": {
                    "type": "number"
                },
                "departureDate": {
                    "type": "string"
                },
                "distributionType": {
                    "type": "string"
                },
                "exchangeDetails": {
                    "$ref": "#/definitions/exchange_issued.ExchangeDetails"
                },
                "flightSelect": {
                    "type": "string"
                },
                "invoiceNumber": {
                    "type": "string"
                },
                "issuanceType": {
                    "$ref": "#/definitions/enum.IssuanceType"
                },
                "logId": {
                    "type": "string"
                },
                "nta": {
                    "type": "number"
                },
                "orderDetailId": {
                    "type": "string"
                },
                "paxes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/issued.PaxRequest"
                    }
                },
                "returnDate": {
                    "type": "string"
                },
                "sessionGroup": {
                    "type": "string"
                },
                "sessionGroupType": {
                    "type": "string"
                },
                "tourCode": {
                    "type": "string"
                },
                "transactionId": {
                    "type": "string"
                }
            }
        },
...
        "issued.IssuedResponse": {
            "type": "object",
            "properties": {
                "additionalData": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "integratorIssuedRequest": {
                    "$ref": "#/definitions/issued.IntegratorIssuedRequest"
                },
                "issuedDetail": {
                    "$ref": "#/definitions/issued.IssuedDetail"
                },
                "mandatoryRequest": {
                    "$ref": "#/definitions/common.MandatoryRequest"
                }
            }
        },
...
    }
}